requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
pytesseract
aiofiles
pdf2image
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
